@@ -1,13 +1,13 @@
 {
   "name": "localize-router",
-  "version": "0.7.0",
-  "description": "An implementation of routes localization for Angular 2",
+  "version": "1.0.0-rc.0",
+  "description": "An implementation of routes localization for Angular",
   "scripts": {
     "test": "karma start",
     "test-watch": "karma start --singleRun=false --autoWatch=true",
     "commit": "npm run prepublish && npm test",
     "prepublish": "ngc && npm run build",
-    "build": "webpack && node copy.js"
+    "build": "webpack"
   },
   "repository": {
     "type": "git",
@@ -33,54 +33,54 @@
   "peerDependencies": {
     "@angular/core": ">=2.0.0",
     "@angular/common": ">=2.0.0",
-    "@angular/http": ">=2.0.0",
     "@angular/router": ">=3.0.0",
     "rxjs": "^5.0.0",
-    "@ngx-translate/core": "^6.0.0"
+    "@ngx-translate/core": ">=6.0.0"
   },
   "devDependencies": {
-    "@angular/common": "~4.1.3",
-    "@angular/compiler": "~4.1.3",
-    "@angular/compiler-cli": "~4.1.3",
-    "@angular/core": "~4.1.3",
-    "@angular/http": "~4.1.3",
-    "@angular/platform-browser": "~4.1.3",
-    "@angular/platform-browser-dynamic": "~4.1.3",
-    "@angular/platform-server": "~4.1.3",
-    "@angular/router": "~4.1.3",
+    "@angular/animations": "~4.3.3",
+    "@angular/common": "~4.3.3",
+    "@angular/compiler": "~4.3.3",
+    "@angular/compiler-cli": "~4.3.3",
+    "@angular/core": "~4.3.3",
+    "@angular/http": "~4.3.3",
+    "@angular/platform-browser": "~4.3.3",
+    "@angular/platform-browser-dynamic": "~4.3.3",
+    "@angular/platform-server": "~4.3.3",
+    "@angular/router": "~4.3.3",
     "@types/es6-promise": "0.0.32",
-    "@types/hammerjs": "2.0.33",
+    "@types/hammerjs": "2.0.34",
     "@types/jasmine": "2.5.47",
-    "@types/node": "~7.0.15",
-    "awesome-typescript-loader": "~3.1.3",
-    "codelyzer": "~3.0.1",
-    "commitizen": "2.9.2",
+    "@types/node": "7.0.22",
+    "awesome-typescript-loader": "3.1.3",
+    "codelyzer": "3.0.1",
+    "commitizen": "2.9.6",
     "concurrently": "3.1.0",
     "core-js": "~2.4.1",
     "cz-conventional-changelog": "1.2.0",
-    "istanbul-instrumenter-loader": "0.2.0",
-    "jasmine-core": "~2.6.1",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "istanbul-instrumenter-loader": "~2.0.0",
+    "jasmine-core": "2.6.2",
+    "karma": "1.7.0",
+    "karma-chrome-launcher": "2.1.1",
     "karma-coverage": "1.1.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-mocha-reporter": "^2.1.0",
-    "karma-remap-coverage": "^0.1.4",
+    "karma-jasmine": "1.1.0",
+    "karma-mocha-reporter": "2.2.3",
+    "karma-remap-coverage": "0.1.4",
     "karma-sourcemap-loader": "0.3.7",
-    "karma-webpack": "~2.0.2",
-    "loader-utils": "0.2.16",
-    "@ngx-translate/core": "~7.0.0",
-    "reflect-metadata": "0.1.8",
-    "rxjs": "^5.4.0",
-    "semantic-release": "4.3.5",
-    "source-map-loader": "0.1.5",
+    "karma-webpack": "2.0.3",
+    "loader-utils": "1.1.0",
+    "@ngx-translate/core": "6.0.0",
+    "reflect-metadata": "0.1.10",
+    "rxjs": "5.4.0",
+    "semantic-release": "6.3.6",
+    "source-map-loader": "0.2.1",
     "ts-helpers": "1.1.2",
-    "tslint": "4.0.2",
-    "tslint-loader": "3.3.0",
-    "typescript": "~2.3.0",
-    "webpack": "~2.4.1",
+    "tslint": "5.3.2",
+    "tslint-loader": "3.5.3",
+    "typescript": "2.3.3",
+    "webpack": "2.6.0",
     "webpack-node-externals": "^1.5.4",
-    "zone.js": "^0.8.12"
+    "zone.js": "0.8.11"
   },
   "config": {
     "commitizen": {
