@@ -59,7 +59,7 @@ module.exports = {
 
   plugins: [
     new webpack.ContextReplacementPlugin(
-      /angular(\\|\/)core(\\|\/)(esm(\\|\/)src|src)(\\|\/)linker/,
+      /angular(\\|\/)core(\\|\/)@angular/,
       helpers.root('./src')
     ),
     new LoaderOptionsPlugin({
@@ -69,8 +69,11 @@ module.exports = {
           emitErrors: false,
           failOnHint: false,
           resourcePath: 'src'
-        },
+        }
       }
     })
-  ]
+  ],
+
+  // disable warnings about bundle size for tests
+  performance: {hints: false}
 };
