@@ -1,10 +1,10 @@
 import { NgModule }             from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+import { RouterModule } from '@angular/router';
 import { LazyComponent } from './lazy/lazy.component';
 import { LocalizeRouterModule } from 'localize-router';
 import { SecondLevelComponent } from "./second-level/second-level.component";
 
-const lazyRoutes: Routes = [
+const lazyRoutes = [
   { path: '',  component: LazyComponent },
   { path: 'secondlevel',  component: SecondLevelComponent }
 ];
