@@ -1,15 +1,29 @@
 import { NgModule }             from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { LocalizeRouterModule } from 'localize-router';
+import { RouterModule } from '@angular/router';
+import { LocalizeRouterModule, LocalizeRouterSettings, LocalizeParser } from 'localize-router';
+import { LocalizeRouterHttpLoader } from 'localize-router-http-loader';
+import { TranslateService } from '@ngx-translate/core';
+import { Location } from '@angular/common';
+import { HttpClient } from '@angular/common/http';
 
-const routes: Routes = [
+export function HttpLoaderFactory(translate: TranslateService, location: Location, settings: LocalizeRouterSettings, http: HttpClient) {
+  return new LocalizeRouterHttpLoader(translate, location, settings, http);
+}
+
+const routes = [
   { path: 'lazy', loadChildren: './+lazy/lazy.module#LazyModule' }
 ];
 
 @NgModule({
   imports: [
     RouterModule.forRoot(routes),
-    LocalizeRouterModule.forRoot(routes)
+    LocalizeRouterModule.forRoot(routes, {
+      parser: {
+        provide: LocalizeParser,
+        useFactory: HttpLoaderFactory,
+        deps: [TranslateService, Location, LocalizeRouterSettings, HttpClient]
+      }
+    })
   ],
   exports: [ RouterModule, LocalizeRouterModule ]
 })
