@@ -1,11 +1,11 @@
 import { NgModule }             from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+import { RouterModule } from '@angular/router';
 import { UsersComponent } from './users/users.component';
 import { UserComponent } from './user/user.component';
 import { ProfileComponent } from './profile/profile.component';
 import { LocalizeRouterModule } from 'localize-router';
 
-const usersRoutes: Routes = [
+const usersRoutes = [
   { path: 'users', children: [
     { path: '', component: UsersComponent },
     { path: ':id', component: UserComponent },
